feat(refacto-connexion-back): add AuthService file et simplify authController
