+++ server/src/controllers/authController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { AppDataSource } from "../data-source";
 import { User } from "../entities/User";
+import { AuthService } from "../service/AuthService";
 import argon2 from "argon2";
 import jwt from "jsonwebtoken";
 
@@ -8,36 +9,17 @@ export const login = async (req: Request, res: Response) => {
   try {
     const { email, password } = req.body;
 
-    const user = await AppDataSource.getRepository(User).findOneBy({ email });
-    if (!user) {
-      return res.status(422).json({ message: "Invalid credentials" });
-    }
+    const { token, user } = await AuthService.login(email, password);
 
-    const isPasswordValid = await argon2.verify(user.password, password);
-
-    if (!isPasswordValid) {
-      return res.status(422).json({ message: "Invalid credentials" });
-    }
-    const token = jwt.sign(
-      { userId: user.id },
-      process.env.JWT_SECRET || "dev-secret",
-      {
-        expiresIn: "1h",
-      }
-    );
-
-    res.status(201).json({
-      message: "Login successful",
+    res.status(200).json({
+      message: "Connexion réussie",
       token,
-      user: {
-        id: user.id,
-        email: user.email,
-        pseudo: user.pseudo,
-        firstname: user.firstname,
-      },
+      user,
     });
-  } catch (error) {
-    console.error("Login error:", error);
-    res.status(500).json({ message: "Internal server error" });
+  } catch (error: any) {
+    console.error("Erreur login", error);
+    res
+      .status(401)
+      .json({ message: error.message || "Identifiants invalides" });
   }
 };
